renamed IDs and rm open ai token
