+++ .github/workflows/reload.yml
@@ -30,5 +30,4 @@ jobs:
           RUNNER_TRACKING_ID: ""
           TOKEN: ${{ secrets.TOKEN }}
           GH_TOKEN: ${{ secrets.GH_TOKEN }}
-          OPENAI_TOKEN: ${{ secrets.OPENAI_TOKEN }}
         run: nohup python main.py &
